+++ src/main/java/za/ac/cput/campusfacilitybooking/service/impl/TimeSlotServiceImpl.java
@@ -8,6 +8,8 @@ import za.ac.cput.campusfacilitybooking.domain.TimeSlot;
 import za.ac.cput.campusfacilitybooking.repository.TimeSlotRepository;
 import za.ac.cput.campusfacilitybooking.service.TimeSlotService;
 
+import java.util.Optional;
+
 @Service
 public class TimeSlotServiceImpl implements TimeSlotService {
 
@@ -24,7 +26,8 @@ public class TimeSlotServiceImpl implements TimeSlotService {
 
     @Override
     public TimeSlot read(String id) {
-        return repository.findById(id).orElse(null);
+        Optional<TimeSlot> timeSlot = repository.findById(id);
+        return timeSlot.orElse(null);
     }
 
     @Override
@@ -34,7 +37,10 @@ public class TimeSlotServiceImpl implements TimeSlotService {
 
     @Override
     public boolean delete(String id) {
-        repository.deleteById(id);
-        return true;
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
